+++ core/src/test/j2cl/entry_point.js
@@ -2,6 +2,8 @@ goog.module('zemeckis.j2cl.BuildTest');
 
 const Zemeckis = goog.require('zemeckis.Zemeckis');
 const ZemeckisConfig = goog.require('zemeckis.ZemeckisConfig');
+const ZemeckisTestUtil = goog.require('zemeckis.ZemeckisTestUtil');
 
 goog.exportSymbol('zemeckis.j2cl.Zemeckis', Zemeckis);
 goog.exportSymbol('zemeckis.j2cl.ZemeckisConfig', ZemeckisConfig);
+goog.exportSymbol('zemeckis.j2cl.ZemeckisTestUtil', ZemeckisTestUtil);