+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Replace the JVM timer thread with a deterministic scheduler controlled through `ZemeckisTestUtil.pumpNext()` and
   `ZemeckisTestUtil.pumpAll()`.
 * Make `ZemeckisTestUtil` available to J2CL tests by removing its dependency on reflection.
+* Patch `rules_closure` to use Protobuf's supported `java_proto_library` implementation.
 
 ### [v0.15](https://github.com/realityforge/zemeckis/tree/v0.15) (2026-07-20) · [Full Changelog](https://github.com/realityforge/zemeckis/compare/v0.14...v0.15)